Modified fitting for minor performance buff
Fitting code was modified to provide minor performance buff. To avoid the constant "Iteration Limit Exceeded" warnings, the warning was temporarily converted to an error, which could then be caught and avoided. Once triggered, the fit step would be re-run with 4x the default max iterations. Additionally, in op_alignAverages_fd, defining the datarange was moved to outside the for loop since it does not change between loop iterations.